Fixed graphics.h abstraction

+++ Engine/src/graphics.cpp
@@ -1,4 +1,4 @@
-#include<graphics.h>
+#include <graphics.h>
 
 #include <iostream>
 
@@ -20,7 +20,7 @@ void Graphics::CreateQuad()
         -0.5f,  0.5f, 0.0f, 1.0f, 1.0f, 0.0f, 0.0f, 1.0f
     };
 
-    float indices[] = {  
+    unsigned int indices[] = {  
         0, 1, 3,  
         1, 2, 3   
     };
